trimVideo: move ffmpeg-download-video -> ffmpeg-download-trimed-video
